MNT: fix LPy compatibility
